Add dimension validation and pretty JSON formatting to useViewPrompt hook
